move mock files
